Adjust TIF saving process
Fixed the issues of not being able to open the TIF images in ImageJ by assuring that saved values are float32
before saving to TIF in the forward projection.
I also created new streamlit page to view the TIF images and metadata.

Included last decimal place of n_medium selected on streamlit page.

All tests pass.

+++ VolumeRaytraceLFM/util.py
@@ -14,5 +14,5 @@ def save_as_tif(file_path, data, metadata):
         if key in metadata['Optical info']:
             del metadata['Optical info'][key]
     with tifffile.TiffWriter(file_path) as tif:
-        tif.save(data, metadata=metadata)
+        tif.save(data.astype('float32'), metadata=metadata)
     return
